+++ packages/core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @under-control/core
 
+## 1.7.21
+
+### Patch Changes
+
+- Fixes again
+
+## 1.7.20
+
+### Patch Changes
+
+- Fixes again
+
+## 1.7.19
+
+### Patch Changes
+
+- Add typings utility, improve typings
+
 ## 1.7.18
 
 ### Patch Changes